+++ src/scite/src/SString.h
@@ -246,7 +246,7 @@ public:
 	}
 	bool startswith(const char *prefix);
 	bool endswith(const char *suffix);
-	int search(const char *sFind, lenpos_t start=0) const;
+	intptr_t search(const char *sFind, lenpos_t start=0) const;
 	bool contains(const char *sFind) const {
 		return search(sFind) >= 0;
 	}
