+++ src/scite/src/StringList.h
@@ -30,7 +30,7 @@ public:
 	//bool InListAbbreviated(const char *s, const char marker);
 	const char *GetNearestWord(const char *wordStart, int searchLen,
 		bool ignoreCase = false, SString wordCharacters="", int wordIndex = -1);
-	char *GetNearestWords(const char *wordStart, int searchLen,
+	char *GetNearestWords(const char *wordStart, size_t searchLen,
 		bool ignoreCase=false, char otherSeparator='\0', bool exactLen=false);
 };
